refactor: move schemas under new github package
