+++ package.json
@@ -6,8 +6,8 @@
     "packages/github_issues",
     "packages/github_pull_requests",
     "packages/github_actions",
-    "packages/schemas",
-    "packages/testing"
+    "packages/testing",
+    "packages/github"
   ],
   "devDependencies": {
     "@anatine/zod-mock": "^3.13.4",
