+++ packages/monthly_roadmap/src/TopFeatureRequests.mjs
@@ -1,6 +1,6 @@
+import { issueSchema } from "github/src/schemas/issues.mjs";
 import { z } from "zod";
 import { formatISOtoLongDate } from "../../date_utils/src/formatter.mjs";
-import { issueSchema } from "../../schemas/src/issue_schema.mjs";
 
 /**
  * Represents a top feature request for reporting.
