+++ packages/monthly_roadmap/src/TopOldest.mjs
@@ -1,7 +1,7 @@
+import { issueSchema } from "github/src/schemas/issues.mjs";
 import { z } from "zod";
 import { diffInDaysFromToday } from "../../date_utils/src/date_diff.mjs";
 import { formatISOtoLongDate } from "../../date_utils/src/formatter.mjs";
-import { issueSchema } from "../../schemas/src/issue_schema.mjs";
 
 /**
  * Represents a top oldest issue for reporting.
