--- packages/schemas/package.json
@@ -1,7 +1,0 @@
-{
-	"name": "schemas",
-	"private": true,
-	"version": "1.0.0",
-	"description": "Zod schemas",
-	"devDependencies": {}
-}
