+++ packages/testing/src/builders/monthly_roadmap.mjs
@@ -1,8 +1,8 @@
+import { issueSchema } from "github/src/schemas/issues.mjs";
 import { TopFeatureRequest } from "../../../monthly_roadmap/src/TopFeatureRequests.mjs";
 import { TopLongRunning } from "../../../monthly_roadmap/src/TopLongRunningPullRequest.mjs";
 import { TopMostCommented } from "../../../monthly_roadmap/src/TopMostCommented.mjs";
 import { TopOldest } from "../../../monthly_roadmap/src/TopOldest.mjs";
-import { issueSchema } from "../../../schemas/src/issue_schema.mjs";
 
 /**
  * Builds an array of top feature requests from GitHub issues
